WIP: Scene management

+++ Firefly/World/Scene/SceneDataModels/Camera.cs
@@ -7,9 +7,9 @@ namespace Firefly.World.Scene.SceneDataModels
   internal class Camera
   {
     public int projectionType { get; set; }
-    public double fov { get; set; }
+    public float fov { get; set; }
     public int farClip { get; set; }
-    public double nearClip { get; set; }
+    public float nearClip { get; set; }
     public List<int> position { get; set; }
     public List<int> rotation { get; set; }
     public string skybox { get; set; }
